test free space check

+++ Other/CameraDemo/checkdiskthread.cpp
@@ -18,11 +18,11 @@ void CheckDiskThread::run()
     int eachpartsleep = sleeptime / 20;
     DiskHelper diskhelper;
     PathHelper pathhelper;
-    diskhelper.setPath(pathhelper.getRecPath());
+    diskhelper.setPath(pathhelper.getRootPath());
     qint64 kbSize = 1024;
     qint64 mbSize = kbSize * 1024;
     qint64 gbSize = mbSize * 1024;
-    qint64 keepBytes = 4 * gbSize;//keep 4Gb free
+    qint64 keepBytes = 2 * gbSize;//keep 2Gb free
     QStringList filters;
     filters << "*." + pathhelper.getFileExtn();
     qDebug() << "check disk thread will start after 20s";
@@ -38,36 +38,48 @@ void CheckDiskThread::run()
     while (_isRunning)
     {
         qint64 free = diskhelper.bytesAvailable();//.bytesFree();
-        if(free < keepBytes)
+        if(free < 0)
         {
-            qDebug() << "no more free space, free space for store: free bytes:" << free;
-            int times = 0;
-            QFileInfoList files = PathHelper::getAllFiles(pathhelper.getRecPath(), filters);
-            while ((!files.isEmpty() && files.count() > 0) && free < keepBytes && times < 100)
+            qDebug() << "warning: bytesAvailable error, switch to bytesFree";
+            free = diskhelper.bytesFree();
+        }
+        if(free < 0)
+        {
+            qDebug() << "disk check free space error";
+        }
+        else
+        {
+            if(free < keepBytes)
             {
-                times++;
-                int index = PathHelper::getEarliestFile(&files);
-                if(index < files.count())
+                qDebug() << "no more free space, free space for store: free bytes:" << free;
+                int times = 0;
+                QFileInfoList files = PathHelper::getAllFiles(pathhelper.getRecPath(), filters);
+                while ((!files.isEmpty() && files.count() > 0) && free < keepBytes && times < 100)
                 {
-                    QFileInfo fileinfo = files.at(index);
-                    QString filename = fileinfo.absoluteFilePath();
-                    if(filename.toLower().contains("temp"))
+                    times++;
+                    int index = PathHelper::getEarliestFile(&files);
+                    if(index < files.count())
                     {
-                        //dot not remove it
-                    }
-                    else
-                    {
-                        QFile file(filename);
-                        if(file.exists())
+                        QFileInfo fileinfo = files.at(index);
+                        QString filename = fileinfo.absoluteFilePath();
+                        if(filename.toLower().contains("temp"))
                         {
-                            file.remove();
+                            //dot not remove it
                         }
-                        qDebug() << times << "remove file:" << filename;
+                        else
+                        {
+                            QFile file(filename);
+                            if(file.exists())
+                            {
+                                file.remove();
+                            }
+                            qDebug() << times << "remove file:" << filename;
+                        }
                     }
+                    files.removeAt(index);
+                    free = diskhelper.bytesAvailable();//.bytesFree();
+                    qDebug() << "now free bytes:" << free;
                 }
-                files.removeAt(index);
-                free = diskhelper.bytesAvailable();//.bytesFree();
-                qDebug() << "now free bytes:" << free;
             }
         }
         for(int i = 0; i < eachpartsleep; i++)
